Add !dnsjedi command
This command retrieves how many chunks are left and returns this.

Retrieval of zabbix item values has been added to the zabbix.py wrapper.
It supports the retrieval of multiple keys at once.

Without any argument the !dnsjedi command gives a summary.
Known arguments are 'left', which returns the number of chunks left per
clustermanager. And 'forecast', which returns the forecast of when the
measurement is predicted to be done.
If other arguments are given the help text is shown.

Remove coloring from matrix.py