Reintroduce zabbix.py, add token auth, bugfixes, fixes #5
Zabbix.py was still used by the zabbix_bot.

Now that MatrixBotApi supports token based logins we actually want to
use it. Update requirements to the latest git commit of the
matrix-bot-api.

Fix a few bugs related to the config migration towards yaml.
Keys got relocated and not propely updated in the code.